♻️ refactor: 헥사고날 아키텍처 전환을 위한 UseCase 리팩토링
Port, Adapter 개념 도입
