+++ application/build.gradle.kts
@@ -1,5 +1,4 @@
 dependencies {
     implementation(project(":domain"))
-    implementation(project(":infrastructure"))
     implementation("org.springframework:spring-context")
 }
