+++ application/src/main/kotlin/io/github/jehwanyoo/redis_1st/repository/port/MovieRepository.kt
@@ -1,4 +1,4 @@
-package io.github.jehwanyoo.redis_1st.repository
+package io.github.jehwanyoo.redis_1st.repository.port
 
 import io.github.jehwanyoo.redis_1st.model.Movie
