+++ application/src/main/kotlin/io/github/jehwanyoo/redis_1st/usecase/port/GetMoviesByTitleContainingUseCase.kt
@@ -1,0 +1,7 @@
+package io.github.jehwanyoo.redis_1st.usecase.port
+
+import io.github.jehwanyoo.redis_1st.model.Movie
+
+interface GetMoviesByTitleContainingUseCase {
+    fun execute(title: String): List<Movie>
+}
