+++ application/src/main/kotlin/io/github/jehwanyoo/redis_1st/usecase/port/GetMoviesUseCase.kt
@@ -1,0 +1,7 @@
+package io.github.jehwanyoo.redis_1st.usecase.port
+
+import io.github.jehwanyoo.redis_1st.model.Movie
+
+interface GetMoviesUseCase {
+    fun execute(): List<Movie>
+}
