+++ core/build.gradle.kts
@@ -1,0 +1,6 @@
+dependencies {
+    implementation("org.springframework.boot:spring-boot-starter")
+    implementation(project(":api"))
+    implementation(project(":application"))
+    implementation(project(":infrastructure"))
+}
