--- domain/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    implementation("org.springframework:spring-context")
-}
