+++ infrastructure/build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
 dependencies {
     implementation("org.springframework:spring-context")
     implementation(project(":domain"))
+    implementation(project(":application"))
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     runtimeOnly("com.mysql:mysql-connector-j")
 }
