+++ infrastructure/src/main/kotlin/io/github/jehwanyoo/redis_1st/repository/entity/MovieEntity.kt
@@ -1,4 +1,4 @@
-package io.github.jehwanyoo.redis_1st.entity
+package io.github.jehwanyoo.redis_1st.repository.entity
 
 import io.github.jehwanyoo.redis_1st.model.Movie
 import jakarta.persistence.*
