+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include("api")
 include("application")
 include("domain")
 include("infrastructure")
+include("core")
